Readme style

+++ thread/pthread/sumall/part_sum.c
@@ -1,5 +1,6 @@
 #include <pthread.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 
 #include "part_sum.h"
@@ -7,14 +8,14 @@
 
 extern int *numbers;         /* this data is shared by the threads      */
 extern int *indices;         /* this data is also shared by the threads */
-extern long long *part_sums; /* this data is also shared by the threads */
+extern int64_t *part_sums; /* this data is also shared by the threads */
 
 /* The thread will execute in this function */
 void *part_sum_runner(void *param) {
 
   int worker_idx = *((int *)param);
 
-  long long sum = 0;
+  int64_t sum = 0;
 
   int begin = indices[worker_idx * 2];
   int end = indices[worker_idx * 2 + 1];
